Taro Api采用继承export的方式，不再复制taro-h5、taro-plugin-react、taro-plugin-vue3里的代码，解决vue框架下缺少的api
